requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
pymongo==4.6.0
pika==1.3.1
requests==2.31.0
python-dotenv==1.0.0
Flask-Caching==2.0.2
google-generativeai==0.1.0

# Optional: Add testing and WSGI server dependencies
pytest==7.4.2
google-auth-oauthlib==1.2.1
redis==5.0.0
